1-basic-api/cdk/requirements.txt
aws-cdk-lib
constructs>=10.0.0,<11.0.0
aws_cdk.aws_apigatewayv2_alpha==2.66.0a0
aws_cdk.aws_apigatewayv2_integrations_alpha==2.66.0a0